refactor code base
