+++ app/src/main/java/com/example/pocketwatching/Apis/Ethplorer/EthplorerClient.java
@@ -1,6 +1,6 @@
 package com.example.pocketwatching.Apis.Ethplorer;
 
-import com.example.pocketwatching.Models.Ethplorer.PortfolioValues.TokenInfo;
+import com.example.pocketwatching.Models.Ethplorer.TokenInfo;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
