+++ app/src/main/java/com/example/pocketwatching/Models/Ethplorer/Price.java
@@ -1,4 +1,4 @@
-package com.example.pocketwatching.Models.Ethplorer.PortfolioValues;
+package com.example.pocketwatching.Models.Ethplorer;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
