+++ app/src/main/java/com/example/pocketwatching/Models/Ethplorer/TxHistory.java
@@ -2,7 +2,6 @@ package com.example.pocketwatching.Models.Ethplorer;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
-import com.google.longrunning.Operation;
 
 import org.json.JSONObject;
 
@@ -11,13 +10,13 @@ import java.util.List;
 public class TxHistory extends JSONObject {
     @SerializedName("operations")
     @Expose
-    private List<com.example.pocketwatching.Models.Ethplorer.PortfolioValues.Operation> operations = null;
+    private List<Operation> operations = null;
 
-    public List<com.example.pocketwatching.Models.Ethplorer.PortfolioValues.Operation> getOperations() {
+    public List<Operation> getOperations() {
         return operations;
     }
 
-    public void setOperations(List<com.example.pocketwatching.Models.Ethplorer.PortfolioValues.Operation> operations) {
+    public void setOperations(List<Operation> operations) {
         this.operations = operations;
     }
 }
